Refactor GM table layout and unify styling
